+++ src/TextShow/GHC/Event.hs
@@ -1,7 +1,8 @@
 {-# LANGUAGE CPP             #-}
 
 #if !defined(__GHCJS__) && !defined(mingw32_HOST_OS) && MIN_VERSION_base(4,4,0)
-{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TemplateHaskell   #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 #endif
 {-|
@@ -29,17 +30,27 @@ module TextShow.GHC.Event (
 # endif
     ) where
 
-import Data.Text.Lazy.Builder (Builder)
+import Data.List (intersperse)
+import Data.Maybe (catMaybes)
+import Data.Monoid.Compat ((<>))
+import Data.Text.Lazy.Builder (Builder, singleton)
 
-import GHC.Event (Event, FdKey)
+import GHC.Event (Event, FdKey, evtRead, evtWrite)
+
+import Language.Haskell.TH.Lib (conT, varE)
 
 import TextShow.Classes (TextShow(showb, showbPrec))
-import TextShow.FromStringTextShow (FromStringShow(..))
+import TextShow.Data.Integral      ()
+import TextShow.System.Posix.Types ()
+import TextShow.TH.Internal (deriveTextShow)
+import TextShow.TH.Names (evtCloseValName, eventIsValName,
+                          fdKeyTypeName, uniqueTypeName, asInt64ValName)
 
 # if MIN_VERSION_base(4,8,1)
 import GHC.Event (Lifetime)
-import TextShow.TH.Internal (deriveTextShow)
 # endif
+
+#include "inline.h"
 
 -- | Convert an 'Event' to a 'Builder'.
 -- This function is only available with @base-4.4.0.0@ or later and is not available
@@ -47,8 +58,15 @@ import TextShow.TH.Internal (deriveTextShow)
 --
 -- /Since: 2/
 showbEvent :: Event -> Builder
-showbEvent = showb . FromStringShow
-{-# INLINE showbEvent #-}
+showbEvent e = singleton '[' <> mconcat (intersperse "," $ catMaybes
+    [ evtRead                 `so` "evtRead"
+    , evtWrite                `so` "evtWrite"
+    , $(varE evtCloseValName) `so` "evtClose"
+    ]) <> singleton ']'
+  where
+    so :: Event -> Builder -> Maybe Builder
+    ev `so` disp | $(varE eventIsValName) e ev = Just disp
+                 | otherwise                   = Nothing
 
 -- | Convert an 'FdKey' to a 'Builder' with the given precedence.
 -- This function is only available with @base-4.4.0.0@ or later and is not available
@@ -56,7 +74,7 @@ showbEvent = showb . FromStringShow
 --
 -- /Since: 2/
 showbFdKeyPrec :: Int -> FdKey -> Builder
-showbFdKeyPrec p = showbPrec p . FromStringShow
+showbFdKeyPrec = showbPrec
 {-# INLINE showbFdKeyPrec #-}
 
 # if MIN_VERSION_base(4,8,1)
@@ -74,9 +92,11 @@ instance TextShow Event where
     showb = showbEvent
     {-# INLINE showb #-}
 
-instance TextShow FdKey where
-    showbPrec = showbFdKeyPrec
-    {-# INLINE showbPrec #-}
+$(deriveTextShow fdKeyTypeName)
+
+instance TextShow $(conT uniqueTypeName) where
+    showb = showb . $(varE asInt64ValName)
+    INLINE_INST_FUN(showb)
 
 # if MIN_VERSION_base(4,8,1)
 $(deriveTextShow ''Lifetime)